day 3 done.
this is so stupid but by filterRows function isn't actually returning the values for me to do my final calculation.

if I log out the values and manually do the math I get the right answer I just can't figure out why oxy_gen and co2_scrub aren't getting set correctly.
I'm running this is quokka so it could just be that